Adds support for JS to configure ring-buffer size.
The default buffer sizes chosen, at least in Win32, are extremely
conservative.  Sure, a skip in audio will be extremely rare, but for any
type of real-time application (video-games), the audio is near-unusable,
because sounds will play several seconds after the event that triggered
them.

I've exposed the buffer-size fields all the way up to the JS api, so
that the application developer can choose the best balance between
reliability and performance to fit their needs.  So far this is only
supported on Win32 (and ignored on other platforms), but I plan on
implementing additional platforms in the future.